Tests/reorganize: move message ID test to own file
